Support dependent option on acts_as_commontator Separated module that automatically builds Commontator Threads
