+++ lib/commontator/acts_as_commontator.rb
@@ -10,15 +10,20 @@ module Commontator::ActsAsCommontator
   module ClassMethods
     def acts_as_commontator(options = {})
       class_exec do
+        association_options = options.extract!(:dependent)
+        association_options[:dependent] ||= :destroy
+
         cattr_accessor :commontator_config
         self.commontator_config = Commontator::CommontatorConfig.new(options)
-        self.is_commontator = true
 
-        has_many :commontator_comments, as: :creator,
-                                        class_name: 'Commontator::Comment'
-        has_many :commontator_subscriptions, as: :subscriber,
-                                             class_name: 'Commontator::Subscription',
-                                             dependent: :destroy
+        has_many :commontator_comments, association_options.merge(
+          as: :creator, class_name: 'Commontator::Comment'
+        )
+        has_many :commontator_subscriptions, association_options.merge(
+          as: :subscriber, class_name: 'Commontator::Subscription'
+        )
+
+        self.is_commontator = true
       end
     end
